media: video: msm: Don't force V4L2 on when using MSM camera
Currently, the primary MSM camera driver does not rely on V4L2.
A V4L2 interface does exist, but is very minimal and is not well tested.
Hence, we add a separate config option for enabling a V4L2 driver interface,
and allow the main MSM camera code to be compiled without V4L2.

Change-Id: Idc01e75a42c1906d8cd4eff757a15b7c0d942f69
Signed-off-by: Dima Zavin <dima@android.com>